requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
requests
python-dotenv==1.0.0
currencyapicom